Added user_provider.dart
